fix: change single producer to randomly select partition server side

+++ src/routing_policy.rs
@@ -8,7 +8,7 @@ use crate::producer::Message;
 pub enum RoutingPolicy {
     #[default]
     RoundRobin,
-    Single(usize),
+    Single,
     Custom(Arc<dyn CustomRoutingPolicy>),
 }
